+++ rodiumai/resources/__init__.py
@@ -1,7 +1,25 @@
 from .audio import Audio
-from .chat import Chat
-from .embeddings import Embeddings
-from .images import Images
-from .video import Video
+from .chat import Chat, ChatNamespace
+from .embeddings import Embeddings, EmbeddingsNamespace
+from .extensions import Extensions
+from .images import Images, ImagesNamespace
+from .messages import Messages, MessagesNamespace
+from .models import Models, ModelsNamespace
+from .video import Video, VideoNamespace
 
-__all__ = ["Chat", "Embeddings", "Images", "Audio", "Video"]
+__all__ = [
+    "Audio",
+    "Chat",
+    "ChatNamespace",
+    "Embeddings",
+    "EmbeddingsNamespace",
+    "Extensions",
+    "Images",
+    "ImagesNamespace",
+    "Messages",
+    "MessagesNamespace",
+    "Models",
+    "ModelsNamespace",
+    "Video",
+    "VideoNamespace",
+]